# Precalculated attribute set hashes