--- a/opentelemetry-sdk/src/attributes/set.rs
+++ b/opentelemetry-sdk/src/attributes/set.rs
@@ -1,3 +1,4 @@
+use std::collections::hash_map::DefaultHasher;
 use std::collections::HashSet;
 use std::{
     cmp::Ordering,
@@ -104,8 +105,8 @@ impl Eq for HashKeyValue {}
 ///
 /// This must implement [Hash], [PartialEq], and [Eq] so it may be used as
 /// HashMap keys and other de-duplication methods.
-#[derive(Clone, Default, Debug, Hash, PartialEq, Eq)]
-pub struct AttributeSet(Vec<HashKeyValue>);
+#[derive(Clone, Default, Debug, PartialEq, Eq)]
+pub struct AttributeSet(Vec<HashKeyValue>, u64);
 
 impl From<&[KeyValue]> for AttributeSet {
     fn from(values: &[KeyValue]) -> Self {
@@ -123,7 +124,12 @@ impl From<&[KeyValue]> for AttributeSet {
             .collect::<Vec<_>>();
         vec.sort_unstable();
 
-        AttributeSet(vec)
+        let mut hasher = DefaultHasher::new();
+        for item in &vec {
+            item.hash(&mut hasher);
+        }
+
+        AttributeSet(vec, hasher.finish())
     }
 }
 
@@ -135,7 +141,12 @@ impl From<&Resource> for AttributeSet {
             .collect::<Vec<_>>();
         vec.sort_unstable();
 
-        AttributeSet(vec)
+        let mut hasher = DefaultHasher::new();
+        for item in &vec {
+            item.hash(&mut hasher);
+        }
+
+        AttributeSet(vec, hasher.finish())
     }
 }
 
@@ -161,5 +172,11 @@ impl AttributeSet {
     /// Iterate over key value pairs in the set
     pub fn iter(&self) -> impl Iterator<Item = (&Key, &Value)> {
         self.0.iter().map(|kv| (&kv.0.key, &kv.0.value))
+    }
+}
+
+impl Hash for AttributeSet {
+    fn hash<H: Hasher>(&self, state: &mut H) {
+        state.write_u64(self.1)
     }
 }