--- a/opentelemetry-sdk/src/attributes/set.rs
+++ b/opentelemetry-sdk/src/attributes/set.rs
@@ -111,7 +111,7 @@ pub struct AttributeSet(Vec<HashKeyValue>, u64);
 impl From<&[KeyValue]> for AttributeSet {
     fn from(values: &[KeyValue]) -> Self {
         let mut seen_keys = HashSet::with_capacity(values.len());
-        let mut vec = values
+        let vec = values
             .iter()
             .rev()
             .filter_map(|kv| {
@@ -122,35 +122,34 @@ impl From<&[KeyValue]> for AttributeSet {
                 }
             })
             .collect::<Vec<_>>();
-        vec.sort_unstable();
 
-        let mut hasher = DefaultHasher::new();
-        for item in &vec {
-            item.hash(&mut hasher);
-        }
-
-        AttributeSet(vec, hasher.finish())
+        AttributeSet::new(vec)
     }
 }
 
 impl From<&Resource> for AttributeSet {
     fn from(values: &Resource) -> Self {
-        let mut vec = values
+        let vec = values
             .iter()
             .map(|(key, value)| HashKeyValue(KeyValue::new(key.clone(), value.clone())))
             .collect::<Vec<_>>();
-        vec.sort_unstable();
 
-        let mut hasher = DefaultHasher::new();
-        for item in &vec {
-            item.hash(&mut hasher);
-        }
-
-        AttributeSet(vec, hasher.finish())
+        AttributeSet::new(vec)
     }
 }
 
 impl AttributeSet {
+    fn new(mut values: Vec<HashKeyValue>) -> Self {
+        values.sort_unstable();
+        let mut hasher = DefaultHasher::new();
+        values.iter().fold(&mut hasher, |mut hasher, item| {
+            item.hash(&mut hasher);
+            hasher
+        });
+
+        AttributeSet(values, hasher.finish())
+    }
+
     /// Returns the number of elements in the set.
     pub fn len(&self) -> usize {
         self.0.len()