@KallDrexx 2023-11-27
Also note that precomputing hashes has the opportunity for further benefits if bounded instruments is implemented, or if #1387 is done.