@lalitb 2023-11-27
Thanks for the PR. Can you also add the benchmark result from main branch (without these changes) for comparison?