@KallDrexx 2023-11-27
> Thanks for the PR. Can you also add the benchmark result from main branch (without these changes) for comparison?

Can you be more specific of which benchmarks you are looking for?  I posted stress test results between `main` and this pr, as well as the `counter_add` benchmarks between main and this pr.