@cijothomas 2023-11-28
https://github.com/open-telemetry/opentelemetry-rust/issues/1405 might be skewing some results.. When we think we use 2 threads due to 2 cpus, we are doing 1 less actually!