@cijothomas 2023-11-28
> > I added the benchmark run I had done previously from main before switching back to the PR for a run.
> 
> Yeah, I meant the benchmark from main. Probably I missed it if it was already there. Thanks :)

benchmarks are probably not the right tool to show gains from this change! (or anything related to improving contentions!) Stress test is proving to be helpful, despite some [limitations](https://github.com/open-telemetry/opentelemetry-rust/issues/1405)!