@KallDrexx 2023-11-28
> #1405 might be skewing some results.. When we think we use 2 threads due to 2 cpus, we are doing 1 less actually!

I had noticed this the other day when I tried setting thread count to 1 and got 0 throughput.  

That being said even if you take 1 away from the `thread count` values in my table, you still get the same results.